refactor: remove old presence engine
